fastApiProject/requirements.txt
alembic==1.13.3
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
asyncpg==0.29.0
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.2
fastapi==0.115.2
greenlet==3.1.1
h11==0.14.0
httptools==0.6.2
idna==3.10
Mako==1.3.5
MarkupSafe==3.0.1
pydantic==2.9.2
pydantic_core==2.23.4
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.40.0
typing_extensions==4.12.2
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
